HIVE-7068 Remove a bunch of unnecessary mapred<->mapreduce indirection.
Remove a bunch of unnecessary code because we were using the Accumulo
mapreduce variants instead of the mapred variants. We still need to
wrap the RangeInputSplit in a FileInputSplit because of some (seemingly
weird) expectation that the InputSplit is a FileInputSplit.

Also removed duplicative configuration by using the RangeInputSplit
state properly. Worked around ACCUMULO-2962 as well by re-constituting
the iterators in the InputSplit in the Mappers.