HIVE-7068 Some tests which show correct range generation given some AST
The hierarchy of the WHERE clauses is very important (e.g. (AND x (OR y z)) as opposed
to (OR z (AND x y)) and the computation of the Ranges to use in accumulo
rely on these to correctly and efficiently prune down the search space for
a hive query.